refactor(rebalancer): remove unnecessary casts

+++ src/Rebalancer/Rebalancer.sol
@@ -303,7 +303,7 @@ contract Rebalancer is Ownable2Step, ReentrancyGuard, ERC165, IOwnershipCallback
 
         if (
             depositData.initiateTimestamp > 0
-                && uint40(block.timestamp) < depositData.initiateTimestamp + _timeLimits.actionCooldown
+                && block.timestamp < depositData.initiateTimestamp + _timeLimits.actionCooldown
         ) {
             // user must wait until the cooldown has elapsed, then call this function to restart the withdrawal process
             revert RebalancerActionCooldown();
@@ -567,11 +567,11 @@ contract Rebalancer is Ownable2Step, ReentrancyGuard, ERC165, IOwnershipCallback
      */
     function _checkValidationTime(uint40 initiateTimestamp) internal view {
         TimeLimits memory timeLimits = _timeLimits;
-        if (uint40(block.timestamp) < initiateTimestamp + timeLimits.validationDelay) {
+        if (block.timestamp < initiateTimestamp + timeLimits.validationDelay) {
             // user must wait until the delay has elapsed
             revert RebalancerValidateTooEarly();
         }
-        if (uint40(block.timestamp) > initiateTimestamp + timeLimits.validationDeadline) {
+        if (block.timestamp > initiateTimestamp + timeLimits.validationDeadline) {
             // user must wait until the cooldown has elapsed, then call `resetDepositAssets` to withdraw the funds
             revert RebalancerActionCooldown();
         }
